refactor(devnet): extract canary devnet lifecycle into MPFS.Devnet (#177)
Relocate the devnet-lifecycle helpers (wallet construction, genesis pre-funding, devnet-server launch) into the moog library so the integration-test harness can reuse them in slice B. No behavior change to the canary.

Tasks: T177-SA
